Throttle RabbitMQ connection in case of error
If you try to declare a queue with wrong settings (for example `durability`), the river creation fails but retry immediately after. It causes a lot of error messages in logs.

We can throttle that and give a delay of 10s before retrying again.

Also, we can add some informations about what is expected by the river as debug level information.

Closes #64.